+++ NEWS.md
@@ -5,6 +5,7 @@
 
 #### RStudio
 
+- Restart commands are now run after restoring the search path + global environment by default. (#14636)
 - The "Soft-wrap R source files" preference now applies to all source files, and has been re-labelled appropriately. (#10940)
 
 #### Posit Workbench